% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/our_colours.R
\name{our_colours}
\alias{our_colours}
\alias{our_colors}
\title{Extract hexcodes for our colours}
\usage{
our_colours(...)

our_colors(...)
}
\description{
@param ... colour names in a comma separated or vector of strings
 @examples {
   our_colours_raw
   our_colours('green', 'blue', 'green')
 }
 @source \url{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
 @export
}
\section{Functions}{
\itemize{
\item \code{our_colors}: Alias for \code{our_colours()}
}}

