% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{our_colours_raw}
\alias{our_colours_raw}
\title{Our colours.}
\format{A named vector of colours.
\describe{
 \item{Names}{Colour names}
 \item{Values}{Hexcodes}
}}
\usage{
our_colours_raw
}
\description{
Our colours.
}
\examples{
 our_colours_raw
}
\keyword{datasets}
