% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/our_colours.R
\name{our_cols.validate}
\alias{our_cols.validate}
\title{Error handling for our_colours}
\usage{
our_cols.validate(colours)
}
\arguments{
\item{colours}{a vector of colour names as strings}
}
\value{
None, will print a message if colour names aren't found in 
our_colours_raw.
}
\description{
Check if the named colours are actually in our dataset, print a warning
if not found.
}
\keyword{internal}
