% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/our_theme.R
\name{our_custom_theme}
\alias{our_custom_theme}
\title{Out custom ggplot 2 theme.}
\usage{
our_custom_theme(base_size = 11, base_family = "sans")
}
\arguments{
\item{base_size}{passed to the selected base theme}

\item{base_family}{passed to the selected base theme}
}
\description{
Choose a base theme from \url{https://ggplot2.tidyverse.org/reference/ggtheme.html}
that's as close as possible to your desired look. Then adjust the custom
elements as required to
}
\keyword{internal}
