% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/our_palettes.R
\name{our_palettes}
\alias{our_palettes}
\title{Access our colour palettes as hexcodes}
\usage{
our_palettes(palette = NA, n = NA, reverse = FALSE)
}
\arguments{
\item{palette}{which palette should be accessed, should match a name from our_palettes_raw}

\item{n}{number of colours to generate from palette}

\item{reverse}{should the order of colours be reversed?}
}
\value{
If palette is NA, return the raw palette data. If n is NA, return 
    the hexcodes of colours in the data, otherwise return n colours interpolated
    from the chosen palette
}
\description{
Access our colour palettes as hexcodes
}
\examples{
 our_palettes()
 our_palettes('default')
 our_palettes('default', reverse = TRUE)
 our_palettes('default', 10)
 our_palettes('default', 2)
}
\seealso{
Other access our colours: \code{\link{our_colours}},
  \code{\link{our_scale}}
}
