% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/our_palettes_interpolator.R
\name{our_palettes_interpolator.validate}
\alias{our_palettes_interpolator.validate}
\title{Error handling for our_palettes_interpolator}
\usage{
our_palettes_interpolator.validate(palette)
}
\arguments{
\item{palette}{string, should match the name of a palette from \code{our_palettes_raw}}
}
\value{
None, will print a message if palette not found.
}
\description{
Check if the named palette is actually in our dataset, print a warning
if not found.
}
\keyword{internal}
