% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{our_palettes_raw}
\alias{our_palettes_raw}
\title{Our palettes.}
\format{A list of colour palettes.
\describe{
 \item{Names}{Palette name}
 \item{Values}{Vector of colour names in this palette. Should match names in \code{our_colours_raw.}}
}}
\usage{
our_palettes_raw
}
\description{
Our palettes.
}
\examples{
 our_palettes_raw
 our_colours(our_palettes_raw[['default']])
}
\keyword{datasets}
