% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/our_scale.R
\name{our_scale_colour_d}
\alias{our_scale_colour_d}
\alias{our_scale_color_d}
\title{Discrete colour scale constructor for our colours}
\usage{
our_scale_colour_d(palette = "default", reverse = F, ...)

our_scale_color_d(palette = "default", reverse = F, ...)
}
\arguments{
\item{palette}{Character name of palette in our_palettes_raw}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to the scale constructor}
}
\description{
Discrete colour scale constructor for our colours
}
\section{Functions}{
\itemize{
\item \code{our_scale_color_d}: Alias for \code{our_scale_colour_d()}
}}

\examples{
\dontrun{
  ggplot(mtcars) + 
    geom_point(aes(x = mpg, 
                   y = wt, 
                   color = as.character(carb))) + 
    our_scale_color_d('hot')
}

}
