% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/our_scale.R
\name{our_scale_fill_c}
\alias{our_scale_fill_c}
\title{Continuous fill scale constructor for our colours}
\usage{
our_scale_fill_c(palette = "default", reverse = F, ...)
}
\arguments{
\item{palette}{Character name of palette in our_palettes_raw}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to the scale constructor}
}
\description{
Continuous fill scale constructor for our colours
}
\examples{
\dontrun{
  ggplot(mtcars) + 
    geom_col(aes(x = reorder(rownames(mtcars), mpg), 
                 y = mpg, 
                 fill = mpg)) +
    coord_flip() +             
    our_scale_fill_c('hot') +
}

}
