% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouRstyle.R
\docType{package}
\name{ouRstyle}
\alias{ouRstyle}
\alias{ouRstyle-package}
\title{ouRstyle: your colours and ggplot2 theme}
\description{
ouRstyle provides a template allowing users to specify colours and colour
palettes which are then built into a package of functions allowing them to
be used in R. A plot theme is also provided that can be customised to reflect
your brand / aesthetic.
}
