% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/our_palettes.R
\name{our_palettes}
\alias{our_palettes}
\title{Access our colour palettes as hexcodes}
\usage{
our_palettes(palette = NA, n = NA, reverse = F)
}
\arguments{
\item{palette}{which palette should be accessed}

\item{n}{number of colours to generate from palette}

\item{reverse}{should the order of colours be reversed?}
}
\value{
If palette is NA, return the raw palette data. If n is NA, return 
    the hexcodes of colours in the data, otherwise return n colours interpolated
    from the chosen palette
}
\description{
Access our colour palettes as hexcodes
}
