% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/our_scale.R
\name{our_scale}
\alias{our_scale}
\title{Scale constructor for our colors}
\source{
Adapted from: \url{https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2}
}
\usage{
our_scale(palette = "default", reverse = F, discrete = T, colour = T,
  ...)
}
\arguments{
\item{palette}{Character name of palette in our_palettes_raw}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{colour}{Boolean indicating whether this is a colour or fill scale}

\item{...}{Additional arguments passed to discrete_scale(),
scale_fill_gradientn(), or scale_colour_gradientn() used
based on discrete and colour_or_fill}
}
\description{
Scale constructor for our colors
}
\keyword{internal}
