% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/our_theme.R
\name{our_theme}
\alias{our_theme}
\title{Our ggplot2 theme}
\usage{
our_theme(base_size = 11, base_family = "sans", horizontal = T)
}
\arguments{
\item{base_size}{passed to base theme (e.g. \code{theme_minimal()})}

\item{base_family}{passed to base theme}

\item{horizontal}{boolean}
}
\description{
Our ggplot2 theme
}
\examples{
{
  ggplot(mtcars) +
  geom_point(aes(x = mpg,
                 y = wt)) +
  facet_grid(vs~cyl) +
  labs(title = 'This is the Title',
       subtitle = 'This is the Subtitle',
       caption = 'This is the caption') +
  our_theme()
}
}
